+++ src/Enums/CallbackDataTypeEnum.php
@@ -6,10 +6,9 @@ use MyCLabs\Enum\Enum;
 
 class CallbackDataTypeEnum extends Enum
 {
-    public const TOAST          = 't';
-    public const START_COMMAND  = 's';
-    public const COMMAND_BUTTON = 'c';
-    public const MENU_SECTION   = 'm';
+    public const TOAST = 't';
+    public const START_COMMAND = 's';
+    public const MENU_SECTION = 'm';
 
     //--------------------------------------------------------------------------
     // These methods are just for IDE autocomplete and not are mandatory.
@@ -22,11 +21,6 @@ class CallbackDataTypeEnum extends Enum
     public static function START_COMMAND()
     {
         return new static (static::START_COMMAND);
-    }
-
-    public static function COMMAND_BUTTON()
-    {
-        return new static (static::COMMAND_BUTTON);
     }
 
     public static function MENU_SECTION()