+++ src/InlineButton/InlineButtonHelpers.php
@@ -3,133 +3,185 @@
 namespace CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\InlineButton;
 
 use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\Enums\InlineButtonTypeEnum;
-use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\InlineButton\Exceptions\BadCallbackDataFormatException;
 use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\InlineButton\Exceptions\TooLongCallbackDataParameterException;
+use Longman\TelegramBot\Entities\InlineKeyboard;
 use Longman\TelegramBot\Entities\InlineKeyboardButton;
-use Longman\TelegramBot\Exception\TelegramException;
 
 trait InlineButtonHelpers
 {
     /**
-     * @param string $_text
+     * Show a popup message.
+     *
+     * @param string $_button_text
      * @param string $_message
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function toast(string $_text, string $_message): InlineKeyboardButton
+    public static function toast(string $_button_text, string $_message): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::TOAST(), $_text, $_message))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::TOAST(), $_button_text, [$_message]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
-     * @param string $_command_name
+     * Start any command.
+     *
+     * @param string      $_button_text    The button text.
+     * @param string      $_command_name   The name of the command which will be executed.
+     * @param string|null $_command_text   The text which will be sent into the command like an usual user text.
+     * @param bool        $_delete_message Delete the message with the inline keyboard after command execution?
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function startCommand(string $_text, string $_command_name): InlineKeyboardButton
-    {
-        return (new InlineButton (InlineButtonTypeEnum::START_COMMAND(), $_text,
-            $_command_name))->getInlineKeyboardButton();
+    public static function startCommand(
+        string $_button_text,
+        string $_command_name,
+        ?string $_command_text = null,
+        bool $_delete_message = false
+    ): InlineKeyboardButton {
+        $data = [
+            $_command_name,
+            $_command_text ?? '',
+            $_delete_message ? CallbackDataEntity::DELETE_MESSAGE_MARKER : ''
+        ];
+
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::START_COMMAND(), $_button_text, $data);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
-     * @param string $_target_command
-     * @param string $_data
+     * Just a short alias for the startCommand() with command text.
+     * (because it works like an usual keyboard button).
+     *
+     * @param string $_button_text    The button text.
+     * @param string $_command_name   The name of the command which will be executed.
+     * @param bool   $_delete_message Delete the message with the inline keyboard after command execution?
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function commandButton(string $_text, string $_target_command, string $_data): InlineKeyboardButton
-    {
-        $data = $_target_command . self::CALLBACK_DATA_DELIMITER . $_data;
-        return (new InlineButton (InlineButtonTypeEnum::COMMAND_BUTTON(), $_text, $data))->getInlineKeyboardButton();
+    public static function button(
+        string $_button_text,
+        string $_command_name,
+        bool $_delete_message = false
+    ): InlineKeyboardButton {
+        return self::startCommand($_button_text, $_command_name, $_button_text, $_delete_message);
     }
 
     /**
-     * @param string $_text
+     * Create a fully equivalent of usual keyboard button.
+     * (The message with inline keyboard will be deleted after button pressing).
+     *
+     * @param string $_button_text  The button text.
+     * @param string $_command_name The name of the command which will be executed.
+     *
+     * @return InlineKeyboardButton
+     * @throws TooLongCallbackDataParameterException
+     */
+    public static function buttonDisposable(string $_button_text, string $_command_name): InlineKeyboardButton
+    {
+        return self::startCommand($_button_text, $_command_name, $_button_text, true);
+    }
+
+    /**
+     * Create an inline keyboard from the string array.
+     *
+     * @param string $_command_name   The name of the command which will be executed.
+     * @param array  $_buttons        The array with keyboard buttons' string labels.
+     * @param bool   $_delete_message Delete the message with the inline keyboard after command execution?
+     *
+     * @return InlineKeyboard
+     */
+    public static function buttonsArray(
+        string $_command_name,
+        array $_buttons,
+        bool $_delete_message = false
+    ): InlineKeyboard {
+        \array_walk_recursive($_buttons, static function (&$value) use ($_command_name, $_delete_message) {
+            $value = self::startCommand($value, $_command_name, $value, $_delete_message);
+        });
+
+        return new InlineKeyboard(...$_buttons);
+    }
+
+    /**
+     * @param string $_button_text
      * @param string $_absolute_path
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function menuSection(string $_text, string $_absolute_path): InlineKeyboardButton
+    public static function menuSection(string $_button_text, string $_absolute_path): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::MENU_SECTION(), $_text,
-            $_absolute_path))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::MENU_SECTION(), $_button_text, [$_absolute_path]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
-     * @param string $_data
+     * @param string $_button_text
+     * @param string $_url
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function url(string $_text, string $_data): InlineKeyboardButton
+    public static function url(string $_button_text, string $_url): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::URL(), $_text, $_data))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::URL(), $_button_text, [$_url]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
+     * @param string $_button_text
      * @param string $_data
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function switchInlineQuery(string $_text, string $_data): InlineKeyboardButton
+    public static function switchInlineQuery(string $_button_text, string $_data): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::SWITCH_INLINE_QUERY(), $_text,
-            $_data))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::SWITCH_INLINE_QUERY(), $_button_text, [$_data]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
+     * @param string $_button_text
      * @param string $_data
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function switchInlineQueryCurrentChat(string $_text, string $_data): InlineKeyboardButton
+    public static function switchInlineQueryCurrentChat(string $_button_text, string $_data): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::SWITCH_INLINE_QUERY_CURRENT_CHAT(), $_text,
-            $_data))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::SWITCH_INLINE_QUERY_CURRENT_CHAT(), $_button_text,
+            [$_data]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
+     * @param string $_button_text
      * @param string $_data
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function callbackGame(string $_text, string $_data): InlineKeyboardButton
+    public static function callbackGame(string $_button_text, string $_data): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::CALLBACK_GAME(), $_text, $_data))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::CALLBACK_GAME(), $_button_text, [$_data]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 
     /**
-     * @param string $_text
+     * @param string $_button_text
      * @param string $_data
+     *
      * @return InlineKeyboardButton
-     * @throws BadCallbackDataFormatException
-     * @throws TelegramException
      * @throws TooLongCallbackDataParameterException
      */
-    public static function pay(string $_text, string $_data): InlineKeyboardButton
+    public static function pay(string $_button_text, string $_data): InlineKeyboardButton
     {
-        return (new InlineButton (InlineButtonTypeEnum::PAY(), $_text, $_data))->getInlineKeyboardButton();
+        $inlineButton = new InlineButton (InlineButtonTypeEnum::PAY(), $_button_text, [$_data]);
+        return $inlineButton->getInlineKeyboardButton();
     }
 }