+++ src/InlineMenuCallbackqueryCommand.php
@@ -8,12 +8,10 @@ use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\InlineButton\Exceptions
 use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\Menu\Exceptions\FullPathWasNotBuiltException;
 use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\Menu\Menu;
 use CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\Telegram\InlineMenuTelegramTrait;
-use Longman\TelegramBot\Commands\Command;
 use Longman\TelegramBot\Commands\SystemCommand;
 use Longman\TelegramBot\Entities\CallbackQuery;
 use Longman\TelegramBot\Entities\InlineKeyboard;
 use Longman\TelegramBot\Entities\ServerResponse;
-use Longman\TelegramBot\Exception\TelegramException;
 
 abstract class InlineMenuCallbackqueryCommand extends SystemCommand
 {
@@ -34,11 +32,13 @@ abstract class InlineMenuCallbackqueryCommand extends SystemCommand
      */
     protected $description = 'Reply to callback query';
 
+    /**
+     * @return Menu
+     */
     abstract protected function getRootMenu(): Menu;
 
     /**
      * @return ServerResponse
-     * @throws TelegramException
      * @throws FullPathWasNotBuiltException
      */
     public function execute(): ServerResponse
@@ -47,7 +47,7 @@ abstract class InlineMenuCallbackqueryCommand extends SystemCommand
 
         // Detect query type.
         try {
-            $query = CallbackDataEntity::fromCallbackQuery($callback_query);
+            $query = CallbackDataEntity::createFromCallbackQuery($callback_query);
         } catch (BadCallbackDataFormatException $e) {
             return $this->answerToast($callback_query, $e->getMessage());
         }
@@ -63,14 +63,17 @@ abstract class InlineMenuCallbackqueryCommand extends SystemCommand
             case CallbackDataTypeEnum::MENU_SECTION():
                 return $this->processMenuSection($query, $callback_query);
 
-            case CallbackDataTypeEnum::COMMAND_BUTTON():
-                return $this->processCommandButton($query, $callback_query);
-
             default:
                 return $this->processDefault($query, $callback_query);
         }
     }
 
+    /**
+     * @param CallbackDataEntity $_query
+     * @param CallbackQuery      $_callback_query
+     *
+     * @return ServerResponse
+     */
     protected function processDefault(CallbackDataEntity $_query, CallbackQuery $_callback_query): ServerResponse
     {
         return $this->answerToast($_callback_query, 'Unknown callback type!');
@@ -78,38 +81,49 @@ abstract class InlineMenuCallbackqueryCommand extends SystemCommand
 
     /**
      * @param CallbackDataEntity $_query
-     * @param CallbackQuery $_callback_query
+     * @param CallbackQuery      $_callback_query
+     *
      * @return ServerResponse
      */
     protected function processToast(CallbackDataEntity $_query, CallbackQuery $_callback_query): ServerResponse
     {
-        return $this->answerToast($_callback_query, $_query->getData());
+        $text = $_query->getData()[0];
+        return $this->answerToast($_callback_query, $text);
     }
 
     /**
      * @param CallbackDataEntity $_query
-     * @param CallbackQuery $_callback_query
+     * @param CallbackQuery      $_callback_query
+     *
      * @return ServerResponse
-     * @throws TelegramException
      */
-    protected function processStartCommand(CallbackDataEntity $_query, CallbackQuery $_callback_query): ServerResponse
-    {
-        $command = $_query->getData();
-        $this->telegram->executeCommandFromCallbackquery($command, $_callback_query);
+    protected function processStartCommand(
+        CallbackDataEntity $_query,
+        CallbackQuery $_callback_query
+    ): ServerResponse {
+        // Get data.
+        [$command, $commandText, $isDeleteMessage] = $_query->getData();
+        $isDeleteMessage = $isDeleteMessage === CallbackDataEntity::DELETE_MESSAGE_MARKER;
+
+        // Delete message if need.
+        $isDeleteMessage && $this->deleteMessage($_callback_query->getMessage());
+
+        // Execute command.
+        $this->telegram->executeCommandFromCallbackquery($command, $commandText, $_callback_query);
         return $this->answerEmpty($_callback_query);
     }
 
     /**
      * @param CallbackDataEntity $_query
-     * @param CallbackQuery $_callback_query
+     * @param CallbackQuery      $_callback_query
+     *
      * @return ServerResponse
      * @throws FullPathWasNotBuiltException
-     * @throws TelegramException
      */
     protected function processMenuSection(CallbackDataEntity $_query, CallbackQuery $_callback_query): ServerResponse
     {
         $rootMenu = $this->getRootMenu();
-        $path     = $_query->getData();
+        $path = $_query->getData()[0];
 
         $childMenu = $rootMenu->findMenuByPath($path);
         if ($childMenu === null) {
@@ -120,48 +134,5 @@ abstract class InlineMenuCallbackqueryCommand extends SystemCommand
         $this->editMessageKeyboard($_callback_query->getMessage(), new InlineKeyboard(...$buttons));
 
         return $this->answerEmpty($_callback_query);
-    }
-
-    /**
-     * @param CallbackDataEntity $_query
-     * @param CallbackQuery $_callback_query
-     * @return ServerResponse
-     */
-    protected function processCommandButton(CallbackDataEntity $_query, CallbackQuery $_callback_query): ServerResponse
-    {
-        $dataCommand = $_query->getExtraDataPart1();
-        $dataPayload = $_query->getExtraDataPart2();
-
-        /** @var Command[] $commands */
-        $commands = \array_filter($this->telegram->getCommandsList(), function ($command) {
-            /** @var Command $command */
-            return !$command->isSystemCommand() && $command->isEnabled() && $command instanceof ProcessInlineButtonInterface;
-        });
-
-        /** @var ProcessInlineButtonInterface $targetCommand */
-        $targetCommand = $this->getCommandWithSomeName($commands, $dataCommand);
-        if ($targetCommand === null) {
-            return $this->answerToast($_callback_query,
-                'Unknown command! Be sure command implements ProcessInlineButtonInterface!');
-        }
-
-        $response = $targetCommand::processInlineButton($dataPayload, $this, $_callback_query);
-        return $response ?? $this->answerEmpty($_callback_query);
-    }
-
-
-    /**
-     * @param Command[] $_commands
-     * @param string $_name
-     * @return Command|null
-     */
-    protected function getCommandWithSomeName(array $_commands, string $_name): ?Command
-    {
-        foreach ($_commands as $command) {
-            if ($command->getName() === $_name) {
-                return $command;
-            }
-        }
-        return null;
     }
 }