+++ src/Utils/PathUtils.php
@@ -4,6 +4,12 @@ namespace CaliforniaMountainSnake\LongmanTelegrambotInlinemenu\Utils;
 
 trait PathUtils
 {
+    /**
+     * @param string $_path
+     * @param string $_delimiter
+     *
+     * @return string
+     */
     protected static function deleteFirstPart(string $_path, string $_delimiter): string
     {
         $ex = \explode($_delimiter, $_path);
@@ -15,6 +21,12 @@ trait PathUtils
         return $_path;
     }
 
+    /**
+     * @param string $_path
+     * @param string $_delimiter
+     *
+     * @return string
+     */
     protected static function getTopPath(string $_path, string $_delimiter): string
     {
         $ex = \explode($_delimiter, $_path);